requirements.txt
autopep8==1.5.1
grpcio==1.27.2
grpcio-tools==1.27.2
protobuf==3.11.3
pycodestyle==2.5.0
pyroute2==0.5.11
six==1.14.0
python-iptables==0.14.0
scapy==2.4.3
bitstring==3.1.6
ntplib==0.3.3
NetfilterQueue==0.8.1
bitarray==1.2.1
